Changed entity component mask to 32 bit

+++ src/ecs/entity.h
@@ -55,9 +55,9 @@ void entity_init(void);
 
 entity_id_t entity_create(void);
 
-bool_t entity_has_components(entity_id_t id, uint16_t comp_mask);
-void entity_set_component(entity_id_t id, uint16_t comp_mask);
-void entity_clear_component(entity_id_t id, uint16_t comp_mask);
+bool_t entity_has_components(entity_id_t id, uint32_t comp_mask);
+void entity_set_component(entity_id_t id, uint32_t comp_mask);
+void entity_clear_component(entity_id_t id, uint32_t comp_mask);
 
 bool_t entity_has_flag(entity_id_t id, uint8_t flag);
 void entity_set_flag(entity_id_t id, uint8_t flag);
